Revert "media: v4l2-core: correctly validate video and metadata ioctls"
This breaks 4k 60fps recording on r9s (specifically with the Samsung Camera app) because the MFC and Camera drivers are set to VFL_TYPE_GRABBER, which was also deprecated in fa83fca55cbeb6a2d0b7e57fe4360429ebc40fd0.
Changing their types to VFL_TYPE_VIDEO did not fix the issue for an unknown reason, so we can instead revert this change to allow the ioctls to work anyway.

This reverts commit 25f576552133fe556217de47ef8917ddf328c6e6.

Signed-off-by: Nahuel Gómez <nahuelgomez329@gmail.com>